Estructura inicial para usuarios

+++ src/main/java/com/example/hexagonal/user/application/command/CreateUserCommand.java
@@ -1,4 +1,5 @@
 package com.example.hexagonal.user.application.command;
 
-public record CreateUserCommand(int id, String username, String password, String role) {
+//DTO
+public record CreateUserCommand(Long id, String username, String password, String role) {
 }
