+++ src/main/java/com/example/hexagonal/user/application/command/UpdateUserCommand.java
@@ -1,4 +1,4 @@
 package com.example.hexagonal.user.application.command;
 
-public record UpdateUserCommand (int id, String username, String password, String role) {
+public record UpdateUserCommand (Long id, String username, String password, String role) {
 }
